Fix cooldowns and expose generic messages
This checkin fixes cooldowns not working. Issue #27.

This checkin also exposes the ability to send and recieve arbitrary
messages.

This is useful if you want to take advantage of the more advanced
features implemented in the protocol, but not yet wrapped by the  SDK.
You can find the protocol document here:
https://dev.mixer.com/reference/interactive/protocol/protocol.pdf