Fix performance issue and set button text and cooldowns
Fixes an performance issue that caused noticeable frame drops for desktop and Unity editor builds when sending outgoing commands.
The ability to set button text. For example:
* MixerInteractive.Button("invincibility").SetText("giveHealth " + "(3)");
The ability to set the spark cost. For example:
* MixerInteractive.Button("giveHealth").SetCost(30);
Includes more of the code as source for easier debugging. Now only a small portion of the code is in DLLs (a small amount of C code necessary to support certain platform-specific features).