Desktop Builds: Fix issue where you didn't have to auth twice for different interactive projects
There was an issue where if you had more than one, different interactive project installed on your PC, the SDK would not re-prompt the user to log in. The correct behavior is to prompt the user to re-authenticate with Mixer the first time when they launch a game with a different interactive project ID.

Behavior will likely not be noticable for developers because you will only have 1 project ID at this point. But the case could hit end users.