Simplify editor UI
As part of troubleshooting an issue over the weekend and in another case, the button in the editor that says "Go Interactive", but doesn't actually do that has caused confusion. In the early days, this feature existed to simulate going interactive, participants, inputs before the service was fully online. However, now that the service is functional the UI is causing confusion.

This change also enables us to move the more commonly used functionality out of the "advanced" section so it will be more easily discoverable. These are features like turning on logging and setting a share code which can benefit from additional visibility.