Round Robin reconnect logic
This change implements round robin reconnect logic. This means if the SDK fails to connect to one websocket host rather than trying the same one, it will pick the next in the list and continue in round robin fashion. This will increase the resliancy of the connection in case one of the hosts can't be connected to.