Allow token to be set for a user
This change allows developers to set the token explicitly for a given user (this is only relevant for Xbox One in the case of more than one user signed in). The SDK will now also pick the CurrentUser by default (instead of the 1st user). And with the new MixerInteractive.Token property, the token can be set to another user's token using the GetTokenAndSignatureAsync API on Xbox.

This change also update the Basic sample to be a toggle instead of a button so it more closely resembles the UI and calling pattern that might appear in a game.