Limit outstanding messages list
The SDK keeps track of messages sent because some need to be replied to / expect replies. There was a problem where on some platforms if the game was run long enough, the size of the list got very big.

This change limits the size of that list so that only messages which need replies are ever added. This will solve the unbounded list issue.