Fix: Create groups
The Unity SDK code would only read the first group and not all the groups. The reason we haven't had reports of this before was that this would only come up if a developer were using the raw SendInteractiveMessage API to create the groups (instead of going through the SDK editor).